backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
beautifulsoup4
slowapi
webcolors
requests
black
python-dotenv